fix(node-fdm): clear the 22 mypy errors the scoped hooks exposed
The previous global mypy hook excluded tests/ and part of src/, so these were
never reported. Scoping mypy per package surfaced them; all four packages are
green now.

Three root causes, none of them a real type bug:

- LayerSpec.config is dict[str, object], so the v18 builder's per-column
  sub-dict copies could not be indexed. Cast each sub-dict once, keeping the
  copy-before-mutate that protects the v17 spec it derives from.
- ODETrainer.mass_encoder is declared nn.Module, but effective_coefficients()
  and b0 live on the concrete encoders. Pass the encoder arguments explicitly
  instead of **common_kwargs (heterogeneous, hence dict[str, object]), and
  narrow at the two call sites. Same narrowing in the smoke test, via a cast
  to MassEncoderLinear.
- PhysicsLayer.set_epoch is a duck-typed hook: bind it through getattr so the
  optional call is expressed in the type system rather than asserted.

Also: drop a duplicate `out` annotation in physics.forward, a stale
type: ignore[override] on MassEncoderLinear.forward, and float() the scalar
PS-efficiency reference return (`**` on an untyped parameter widens to Any).

test_identifiability_diag.py goes to archive/ with the script it imports.
Full node-fdm suite green: 442 passed, 3 skipped.

+++ packages/node-fdm/src/node_fdm/layers/ps_efficiency.py
@@ -160,7 +160,8 @@ def eta_o_reference(
         arg = ratio - 1.0
         h_0 = (1.0 - _H0_CURVATURE * arg * arg) * (1.0 + omega * arg * arg)
 
-    return h_0 * eta_o_b
+    # h_1 comes from `**` on an untyped parameter, so the product widens to Any.
+    return float(h_0 * eta_o_b)
 
 
 class PSEfficiencyLayer(nn.Module):
